feat(web): add game context provider
